Restructured project and renames files
